ingest arrays refactored
