+++ src/main/java/org/broadinstitute/hellbender/tools/variantdb/SchemaUtils.java
@@ -48,7 +48,8 @@ public class SchemaUtils {
     public static final long chromAdjustment = 1000000000000L;
 
     public static long encodeLocation(String chrom, int position) {
-        int chromosomeIndex = ChromosomeEnum.valueOfContig(chrom).index;
+        ChromosomeEnum val = ChromosomeEnum.valueOfContig(chrom);
+        int chromosomeIndex = val.index;
         long adjustedLocation = Long.valueOf(chromosomeIndex) * chromAdjustment + Long.valueOf(position);
         return adjustedLocation;
     }
